pyproject.toml
[tool.poetry]
name = "first-etl"
version = "0.1.0"
description = ""
authors = ["gussfer <oliveira.gustavo@unifesp.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.122.0"
duckdb = "^0.10.0"
streamlit = "^1.32.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.28"
gdown = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
